feat: use new domain service for V1 API
